refactor(web): ディレクトリ規約と Props 命名に合わせて分析 UI を整理
