+++ web/src/app/dashboard/components/analytics/UsageSection.tsx
@@ -8,18 +8,14 @@ import {
   YAxis,
 } from "recharts";
 import { useUsageAnalytics } from "~/app/dashboard/hooks/useAnalytics";
-import {
-  AXIS_STYLE,
-  getColorAt,
-  TOOLTIP_STYLE,
-} from "~/components/chat/tool-uis/chart-helpers";
+import { AXIS_STYLE, getColorAt, TOOLTIP_STYLE } from "~/lib/chart-helpers";
 import {
   SectionCard,
   SectionEmpty,
   SectionError,
   SectionLoading,
 } from "./SectionCard";
-import { formatCostUsd, pivotWeeklyUsage } from "./usage-helpers";
+import { formatCostUsd, pivotWeeklyUsage } from "./helpers";
 
 export const UsageSection = () => {
   const { data, isLoading, error } = useUsageAnalytics(12);
